moved implementation of Settings from game to internal/game
