+++ mocks/gameRunner.go
@@ -2,7 +2,8 @@ package mocks
 
 import (
 	"github.com/GomeBox/gome/adapters"
-	"github.com/GomeBox/gome/internal/game"
+	"github.com/GomeBox/gome/game"
+	internalGame "github.com/GomeBox/gome/internal/game"
 )
 
 type GameRunnerMock struct {
@@ -10,11 +11,11 @@ type GameRunnerMock struct {
 	InitializeCnt int
 }
 
-func (runner *GameRunnerMock) Loop(initialize game.InitializeCallback, update game.UpdateCallback, draw game.DrawCallback) error {
+func (runner *GameRunnerMock) Loop(initialize internalGame.InitializeCallback, update internalGame.UpdateCallback, draw internalGame.DrawCallback) error {
 	return nil
 }
 
-func (runner *GameRunnerMock) Init(createAdapters game.CreateAdapters, settings game.Settings) error {
+func (runner *GameRunnerMock) Init(createAdapters internalGame.CreateAdapters, settings game.Settings) error {
 	runner.InitializeCnt++
 	return nil
 }
